# backend/src/main/java/io/reliza/ws/tea/ComponentsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package io.reliza.ws.tea;

import org.springframework.lang.Nullable;
import io.reliza.model.tea.TeaIdentifierType;
import io.reliza.model.tea.TeaPaginatedComponentResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-29T10:44:15.267909500-04:00[America/Toronto]", comments = "Generator version: 7.21.0")
@Validated
@Tag(name = "TEA Component", description = "the TEA Component API")
public interface ComponentsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_QUERY_TEA_COMPONENTS = "/components";
    /**
     * GET /components
     * Returns a list of TEA components. Note that multiple components may match.
     *
     * @param pageOffset Pagination offset (optional, default to 0)
     * @param pageSize Pagination offset (optional, default to 100)
     * @param idType Type of identifier specified in the &#x60;idValue&#x60; parameter (optional)
     * @param idValue If present, only the objects with the given identifier value will be returned. (optional)
     * @return A paginated response containing TEA Components (status code 200)
     *         or Request was Invalid (status code 400)
     */
    @Operation(
        operationId = "queryTeaComponents",
        description = "Returns a list of TEA components. Note that multiple components may match.",
        tags = { "TEA Component" },
        responses = {
            @ApiResponse(responseCode = "200", description = "A paginated response containing TEA Components", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TeaPaginatedComponentResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Request was Invalid")
        },
        security = {
            @SecurityRequirement(name = "basicAuth"),
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ComponentsApi.PATH_QUERY_TEA_COMPONENTS,
        produces = { "application/json" }
    )
    default ResponseEntity<TeaPaginatedComponentResponse> queryTeaComponents(
        @Parameter(name = "pageOffset", description = "Pagination offset", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageOffset", required = false, defaultValue = "0") Long pageOffset,
        @Parameter(name = "pageSize", description = "Pagination offset", in = ParameterIn.QUERY) @Valid @RequestParam(value = "pageSize", required = false, defaultValue = "100") Long pageSize,
        @Parameter(name = "idType", description = "Type of identifier specified in the `idValue` parameter", in = ParameterIn.QUERY) @Valid @RequestParam(value = "idType", required = false) @Nullable TeaIdentifierType idType,
        @Parameter(name = "idValue", description = "If present, only the objects with the given identifier value will be returned.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "idValue", required = false) @Nullable String idValue
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"totalResults\" : 1, \"pageStartIndex\" : 0, \"pageSize\" : 6, \"results\" : [ { \"identifiers\" : [ { \"idType\" : \"\", \"idValue\" : \"idValue\" }, { \"idType\" : \"\", \"idValue\" : \"idValue\" } ], \"name\" : \"name\", \"uuid\" : \"\" }, { \"identifiers\" : [ { \"idType\" : \"\", \"idValue\" : \"idValue\" }, { \"idType\" : \"\", \"idValue\" : \"idValue\" } ], \"name\" : \"name\", \"uuid\" : \"\" } ], \"timestamp\" : \"2024-03-20T15:30:00Z\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}