# tea-spec/openapi.yaml
...
openapi: 3.1.1
...
info:
  title: Transparency Exchange API
  summary: The OWASP Transparency Exchange API specification for consumers and publishers
  description: TBC
  contact:
    name: TEA Working Group
    email: tbc@somewhere.tld
    url: https://github.com/CycloneDX/transparency-exchange-api
  license:
    name: Apache 2.0
    url: https://github.com/CycloneDX/transparency-exchange-api/blob/main/LICENSE
  version: 0.4.0
servers:
  - url: http://localhost/tea/v1
    description: Local development
paths:
...
  /component/{uuid}:
    get:
      description: Get a TEA Component
      operationId: getTeaComponentById
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Component in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
      responses:
        '200':
          description: Requested TEA Component found and returned
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/component"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Component
...
  /components:
    get:
      description: Returns a list of TEA components. Note that multiple components may match.
      operationId: queryTeaComponents
      parameters:
        - $ref: "#/components/parameters/page-offset"
        - $ref: "#/components/parameters/page-size"
        - $ref: "#/components/parameters/id-type"
        - $ref: "#/components/parameters/id-value"
      responses:
        '200':
          $ref: "#/components/responses/paginated-component"
        '400':
          $ref: "#/components/responses/400-invalid-request"
      tags:
        - TEA Component
...
  /componentRelease/{uuid}:
    get:
      description: Get the TEA Component Release with its latest collection
      operationId: getComponentReleaseById
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Component Release in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
      responses:
        '200':
          description: Requested TEA Component Release and its latest Collection found and returned
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/component-release-with-collection"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Component Release
  /componentRelease/{uuid}/cle:
    get:
      description: Get the CLE (Common Lifecycle Enumeration) data for a TEA Component Release
      operationId: getCleByComponentReleaseId
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Component Release in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
      responses:
        '200':
          description: CLE data for the requested TEA Component Release found and returned
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/cle"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - CLE
  /componentRelease/{uuid}/collection/latest:
    get:
      description: Get the latest TEA Collection belonging to the TEA Component Release
      operationId: getLatestCollection
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Component Release in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
      responses:
        '200':
          description: Requested TEA Collection found and returned
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/collection"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Component Release
...
  /componentRelease/{uuid}/collections:
    get:
      description: Get the TEA Collections belonging to the TEA Component Release
      operationId: getCollectionsByReleaseId
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Component Release in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
      responses:
        '200':
          description: Requested TEA Collection found and returned
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/collection"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Component Release
...
  /componentRelease/{uuid}/collection/{collectionVersion}:
    get:
      description: Get a specific Collection (by version) for a TEA Component Release by its UUID
      operationId: getCollection
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of TEA Collection in the TEA server
          schema:
            "$ref": "#/components/schemas/uuid"
        - name: collectionVersion
          in: path
          required: true
          description: Version of TEA Collection
          schema:
            type: integer
      responses:
        '200':
          description: Requested TEA Collection Version found and returned
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/collection"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Component Release
...
components:
  schemas:
...
    date-time:
      type: string
      description: Timestamp
      format: date-time
      pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
      example: '2024-03-20T15:30:00Z'
    identifier:
      type: object
      description: An identifier with a specified type
      properties:
        idType:
          description: Type of identifier, e.g. `TEI`, `PURL`, `CPE`
          "$ref": "#/components/schemas/identifier-type"
        idValue:
          description: Identifier value
          type: string
    identifier-type:
      type: string
      description: Enumeration of identifiers types
      enum:
        - CPE
        - TEI
        - PURL
        - COMPLIANCE_DOCUMENT
...
    uuid:
      type: string
      description: A UUID
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"

    #
    # TEA Product
    #
...
    component:
      type: object
      description: A TEA component
      properties:
        uuid:
          description: A unique identifier for the TEA component
          "$ref": "#/components/schemas/uuid"
        name:
          type: string
          description: Component name
        identifiers:
          type: array
          description: List of identifiers for the component
          items:
            "$ref": "#/components/schemas/identifier"
      required:
        - uuid
        - name
        - identifiers
      examples:
        - uuid: 3910e0fd-aff4-48d6-b75f-8bf6b84687f0
          name: Apache Log4j API
          identifiers:
            - idType: PURL
              idValue: pkg:maven/org.apache.logging.log4j/log4j-api
        - uuid: b844c9bd-55d6-478c-af59-954a932b6ad3
          name: Apache Log4j Core
          identifiers:
            - idType: CPE
              idValue: cpe:2.3:a:apache:log4j
            - idType: PURL
              idValue: pkg:maven/org.apache.logging.log4j/log4j-core

#
    # Reference to a component, in some cases directly to a specific release
    #
    # The release reference (release UUID) is only used in cases where a product
    # name includes a version and this version of the product always includes
    # the same releases of the component.
...
    release:
      type: object
      description: A TEA Component Release
      properties:
        uuid:
          description: A unique identifier for the TEA Component Release
          "$ref": "#/components/schemas/uuid"
        component:
          description: UUID of the TEA Component this release belongs to
          "$ref": "#/components/schemas/uuid"
        componentName:
          description: Name of the TEA Component this release belongs to
          type: string
          example: tomcat
        version:
          description: Version number
          type: string
          example: 1.2.3
        createdDate:
          description: Timestamp when this Release was created in TEA (for sorting purposes)
          "$ref": "#/components/schemas/date-time"
        releaseDate:
          description: Timestamp of the release
          "$ref": "#/components/schemas/date-time"
        preRelease:
          type: boolean
          description: |
            A flag indicating pre-release (or beta) status.
            May be disabled after the creation of the release object, but can't be enabled after creation of an object.
        identifiers:
          type: array
          description: List of identifiers for the component
          items:
            "$ref": "#/components/schemas/identifier"
        distributions:
          type: array
          description: List of different formats of this component release
          items:
            "$ref": "#/components/schemas/release-distribution"
      required:
        - uuid
        - version
        - createdDate
      examples:
        # Apache Tomcat 11.0.7
        - uuid: 605d0ecb-1057-40e4-9abf-c400b10f0345
          version: "11.0.7"
          createdDate: 2025-05-07T18:08:00Z
          releaseDate: 2025-05-12T18:08:00Z
          identifiers:
            - idType: PURL
              idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.7
          distributions:
            - distributionId: 6a0d58e1-4896-4f1e-83f7-5feb6c032537
              description: Core binary distribution, zip archive
              identifiers:
                - idType: PURL
                  idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.6?type=zip
              checksums:
                - algType: SHA_256
                  algValue: 9da736a1cdd27231e70187cbc67398d29ca0b714f885e7032da9f1fb247693c1
              url: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.zip
              signatureUrl: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.zip.asc
            - distributionId: dba01c13-dd96-4928-be72-9c87ffa8cab8
              description: Core binary distribution, tar.gz archive
              identifiers:
                - idType: PURL
                  idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.6?type=tar.gz
              checksums:
                - algType: SHA_256
                  algValue: 2fcece641c62ba1f28e1d7b257493151fc44f161fb391015ee6a95fa71632fb9
              url: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.tar.gz
              signatureUrl: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.tar.gz.asc
            - distributionId: cfe068c2-fae7-43d0-97ec-5ea092454040
              description: Core binary distribution, Windows x64 zip archive
              identifiers:
                - idType: PURL
                  idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.6?classifier=windows-x64&type=zip
              checksums:
                - algType: SHA_256
                  algValue: 62a5c358d87a8ef21d7ec1b3b63c9bbb577453dda9c00cbb522b16cee6c23fc4
              url: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6-windows-x64.zip
              signatureUrl: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.6.zip.asc
            - distributionId: de45ffaf-e4b5-47b5-be28-444a76df098e
              description: Core binary distribution, Windows Service Installer (MSI)
              checksums:
                - algType: SHA_512
                  algValue: 1d3824e7643c8aba455ab0bd9e67b14a60f2aaa6aa7775116bce40eb0579e8ced162a4f828051d3b867e96ee2858ec5da0cc654e83a83ba30823cbea0df4ff96
              url: https://dlcdn.apache.org/tomcat/tomcat-11/v11.0.7/bin/apache-tomcat-11.0.7.exe
              signatureUrl: https://downloads.apache.org/tomcat/tomcat-11/v11.0.7/bin/apache-tomcat-11.0.7.exe.asc
        # A pre-release of Apache Tomcat
        - uuid: 95f481df-f760-47f4-b2f2-f8b76d858450
          version: "11.0.0-M26"
          createdDate: 2024-09-13T17:49:00Z
          preRelease: true
          identifiers:
            - idType: PURL
              idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.0-M26
...
    component-release-with-collection:
      type: object
      description: A TEA Component Release combined with its latest collection
      properties:
        release:
          description: The TEA Component Release information
          $ref: "#/components/schemas/release"
        latestCollection:
          description: The latest TEA Collection for this component release
          $ref: "#/components/schemas/collection"
      required:
        - release
        - latestCollection
      examples:
        - release:
            uuid: 605d0ecb-1057-40e4-9abf-c400b10f0345
            version: "11.0.7"
            createdDate: 2025-05-07T18:08:00Z
            releaseDate: 2025-05-12T18:08:00Z
            identifiers:
              - idType: PURL
                idValue: pkg:maven/org.apache.tomcat/tomcat@11.0.7
          latestCollection:
            uuid: 605d0ecb-1057-40e4-9abf-c400b10f0345
            version: 2
            date: 2025-05-12T18:08:00Z
            belongsTo: COMPONENT_RELEASE
            updateReason:
              type: INITIAL_RELEASE
              comment: Initial collection for this release
            artifacts:
              - uuid: 1cb47b95-8bf8-3bad-a5a4-0d54d86e10ce
                name: Build SBOM
                type: BOM
                formats:
                  - mediaType: application/vnd.cyclonedx+xml
                    description: CycloneDX SBOM (XML)
                    url: https://repo.maven.apache.org/maven2/org/apache/tomcat/tomcat/11.0.7/tomcat-11.0.7-cyclonedx.xml
                    checksums:
                      - algType: SHA-256
                        algValue: 9da736a1cdd27231e70187cbc67398d29ca0b714f885e7032da9f1fb247693c1
              - uuid: dfa35519-9734-4259-bba1-3e825cf4be06
                name: Vulnerability Disclosure Report
                type: VULNERABILITIES
                formats:
                  - mediaType: application/vnd.cyclonedx+xml
                    description: CycloneDX VDR (XML)
                    url: https://tomcat.apache.org/cyclonedx/vdr.xml
                    checksums:
                      - algType: SHA-256
                        algValue: 75b81020b3917cb682b1a7605ade431e062f7a4c01a412f0b87543b6e995ad2a

    #
    # TEA Collection and related objects
    #
    collection:
      type: object
      description: A collection of security-related documents
      properties:
        uuid:
          description: |
            UUID of the TEA Collection object.
            This matches the UUID of the associated TEA Component Release or TEA Product Release object.
            When updating a collection, only the `version` is changed.
          "$ref": "#/components/schemas/uuid"
        version:
          type: integer
          description: |
            TEA Collection version, incremented each time its content changes.
            Versions start with 1.
        date:
          description: The date when the TEA Collection version was created.
          "$ref": "#/components/schemas/date-time"
        belongsTo:
          description: Indicates whether this collection belongs to a Component Release or a Product Release
          "$ref": "#/components/schemas/collection-belongs-to-type"
        updateReason:
          description: Reason for the update/release of the TEA Collection object.
          "$ref": "#/components/schemas/collection-update-reason"
        artifacts:
          type: array
          description: List of TEA Artifact objects.
          items:
            "$ref": "#/components/schemas/artifact"
      examples:
        # Documents in the latest release of Log4j Core
        - uuid: 4c72fe22-9d83-4c2f-8eba-d6db484f32c8
          version: 10
          date: 2024-12-13T00:00:00Z
          updateReason:
            type: ARTIFACT_UPDATED
            comment: VDR file updated
          artifacts:
            - uuid: 1cb47b95-8bf8-3bad-a5a4-0d54d86e10ce
              name: Build SBOM
              type: BOM
              formats:
                - mediaType: application/vnd.cyclonedx+xml
                  description: CycloneDX SBOM (XML)
                  url: https://repo.maven.apache.org/maven2/org/apache/logging/log4j/log4j-core/2.24.3/log4j-core-2.24.3-cyclonedx.xml
                  signatureUrl: https://repo.maven.apache.org/maven2/org/apache/logging/log4j/log4j-core/2.24.3/log4j-core-2.24.3-cyclonedx.xml.asc
                  checksums:
                    - algType: MD5
                      algValue: 2e1a525afc81b0a8ecff114b8b743de9
                    - algType: SHA-1
                      algValue: 5a7d4caef63c5c5ccdf07c39337323529eb5a770
            - uuid: dfa35519-9734-4259-bba1-3e825cf4be06
              version: 7
              name: Vulnerability Disclosure Report
              type: VULNERABILITIES
              formats:
                - mediaType: application/vnd.cyclonedx+xml
                  description: CycloneDX VDR (XML)
                  url: https://logging.apache.org/cyclonedx/vdr.xml
                  checksums:
                    - algType: SHA-256
                      algValue: 75b81020b3917cb682b1a7605ade431e062f7a4c01a412f0b87543b6e995ad2a
...
    unknown-error-type:
      type: string
      description: Classification of TEA error response
      enum:
        - OBJECT_UNKNOWN
        - OBJECT_NOT_SHAREABLE
    #
    # Types used in API responses
    #
    error-response:
      type: object
      description: Error response
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/unknown-error-type"
      required:
        - error
...
    cle-definitions:
      type: object
      description: Container for reusable CLE policy definitions
      properties:
        support:
          type: array
          description: List of support policies
          items:
            $ref: "#/components/schemas/cle-support-definition"
    cle:
      type: object
      description: |
        Common Lifecycle Enumeration (CLE) object based on ECMA-428 TC54 TG3 CLE Specification v1.0.0.
        Contains lifecycle events and optional reusable definitions for a component or product.
      properties:
        events:
          type: array
          description: |
            Ordered array of CLE Event objects representing lifecycle events.
            MUST be ordered by ID in descending order (newest events with highest IDs first).
          items:
            $ref: "#/components/schemas/cle-event"
        definitions:
          description: Container for reusable policy definitions referenced by events
          $ref: "#/components/schemas/cle-definitions"
      required:
        - events
      examples:
        - events:
            - id: 3
              type: endOfSupport
              effective: "2025-06-01T00:00:00Z"
              versions:
                - range: "vers:npm/>=1.0.0|<2.0.0"
              supportId: standard
              published: "2025-01-01T00:00:00Z"
            - id: 2
              type: endOfDevelopment
              effective: "2025-01-01T00:00:00Z"
              versions:
                - version: "1.0.0"
              supportId: standard
              published: "2024-06-01T00:00:00Z"
            - id: 1
              type: released
              effective: "2024-01-01T00:00:00Z"
              version: "1.0.0"
              license: Apache-2.0
              published: "2024-01-01T00:00:00Z"
          definitions:
            support:
              - id: standard
                description: Standard product support policy
                url: https://example.com/support/standard
...
    paginated-component-response:
      type: object
      description: A paginated response containing TEA Components
      allOf:
        - $ref: "#/components/schemas/pagination-details"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/component"
...
  responses:
...
    400-invalid-request:
      description: Request was Invalid
      content:
        application/json: {}
...
    404-object-by-id-not-found:
      description: Object requested by identifier not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error-response"
...
    paginated-component:
      description: A paginated response containing TEA Components
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/paginated-component-response"
...
  parameters:
...
    page-offset:
      name: pageOffset
      description: Pagination offset
      in: query
      required: false
      schema:
        type: integer
        format: int64
        default: 0
    page-size:
      name: pageSize
      description: Pagination offset
      in: query
      required: false
      schema:
        type: integer
        format: int64
        default: 100
    #
    # Query by identifier
    #
    # Since OpenAPI 3.0 it is possible to use RFC 6570-based serialization for JSON parameters of type array or object:
    #   https://swagger.io/docs/specification/v3_0/serialization/
    #
    # Unfortunately many tools don't support it, for example,
    # the `openapi-generator` for Java does not handle this correctly.
    #   https://github.com/OpenAPITools/openapi-generator/issues/4808
    #
    # This can be uncommented, when RFC 6570-base serialization reaches a wider adoption:
    #
    # identifier-param:
    #   name: identifierParam
    #   description: If present, only the objects with the given identifier will be returned.
    #   in: query
    #   schema:
    #     $ref: "#/components/schemas/identifier"
    #   style: form
    #   explode: true
    #
    # In the meantime we explode the object manually:
    id-type:
      # To allow RFC 6570 in the future without breaking changes to the HTTP API,
      # the name of this parameter should be identical to the equivalent property in /components/schemas/identifier
      name: idType
      description: Type of identifier specified in the `idValue` parameter
      in: query
      schema:
        $ref: "#/components/schemas/identifier-type"
    id-value:
      # To allow RFC 6570 in the future without breaking changes to the HTTP API,
      # the name of this parameter should be identical to the equivalent property in /components/schemas/identifier
      name: idValue
      description: If present, only the objects with the given identifier value will be returned.
      in: query
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    basicAuth:
      type: http
      scheme: basic
security:
  - bearerAuth: []
  - basicAuth: []
...